Create transaction-actitivy component
