+++ ui/app/components/ui/popover/popover.stories.js
@@ -2,6 +2,7 @@ import React from 'react'
 import { storiesOf } from '@storybook/react'
 import PopOver from './popover.component'
 import TransactionDetails from '../transaction-details'
+import TransactionActivity from '../transaction-activity'
 import { text } from '@storybook/addon-knobs/react'
 
 import { Tab, Tabs, TabList, TabPanel } from 'react-tabs'
@@ -18,15 +19,6 @@ const close = e => {
   e.preventDefault()
   console.log('close window')
 }
-
-const activities = [
-  'Transaction created with a value of 0.000023 ETH at 09:50 on 8/8/2019.',
-  'Transaction submitted with gas fee of 0 WEI at 09:47 on 8/8/2019.',
-  'Transaction confirmed at 09:45 on 8/8/2019.',
-  'Transaction created with a value of 0.000023 ETH at 09:50 on 8/8/2019.',
-  'Transaction submitted with gas fee of 0 WEI at 09:47 on 8/8/2019.',
-  'Transaction confirmed at 09:45 on 8/8/2019.',
-]
 
 const Fees = () => (
   <div className="fees">
@@ -54,32 +46,6 @@ const Fees = () => (
         <span className="eth">0.00567 ETH</span>
       </p>
     </div>
-  </div>
-)
-
-const Activity = () => (
-  <div className="activity">
-    <div className="header">
-      <a href="#" onClick={e => e.preventDefault()} className="etherscan">
-        View on Etherscan
-      </a>
-      <h3>Activity</h3>
-    </div>
-    <ul>
-      {
-        activities.map((item, index) => (
-          <li key={index}>
-            <div className="line" />
-            <div className="bullet">
-              <svg width="7" height="7" viewBox="0 0 7 7" fill="none" xmlns="http://www.w3.org/2000/svg">
-                <circle cx="3.5" cy="3.5" r="3.5" fill="#037DD6" />
-              </svg>
-            </div>
-            <span>{item}</span>
-          </li>
-        ))
-      }
-    </ul>
   </div>
 )
 
@@ -123,7 +89,7 @@ storiesOf('PopOver', module)
             </div>
           </div>
           <TransactionDetails amount={0.00} fee={0.04} ethCost={0.000185} />
-          <Activity />
+          <TransactionActivity />
         </div>
       </PopOver>
     </div>
