+++ ui/app/components/ui/transaction-activity/index.js
@@ -1,0 +1,3 @@
+import TransactionActivity from './transaction-activity.component'
+
+export default TransactionActivity
